+++ wiki/Changelog.md
@@ -13,6 +13,7 @@ Configuration
   * New option ShowTabBar (#944).
   * New command-line option --tabbar (#944).
   * New user-definable functions clear-title, refresh.
+  * Option KeyFunctions: flexible specification of modifiers (in any order).
 
 ### 3.3.0 (6 August 2020) ###